% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadagno.r
\name{extract_date}
\alias{extract_date}
\title{EXTRACT DATE
description: extract date information from the string input
PARAMETERS}
\usage{
extract_date(s)
}
\arguments{
\item{s}{character string

OUTPUT}
}
\value{
character date object
}
\description{
EXTRACT DATE
description: extract date information from the string input
PARAMETERS
}
