% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadagno.r
\name{get_month}
\alias{get_month}
\title{GET MONTH
description: identify and translate the month name in x
PARAMETERS}
\usage{
get_month(x)
}
\arguments{
\item{x}{character vector

OUTPUT}
}
\value{
x with english month name
}
\description{
GET MONTH
description: identify and translate the month name in x
PARAMETERS
}
