% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadagno.r
\name{get_precision}
\alias{get_precision}
\title{#############################################################
GET PRECISION
DESCRIPTION
returns the precision (number of digits) after the comma for
the variable of interest
PARAMETERS}
\usage{
get_precision(var)
}
\arguments{
\item{var}{variable (as defined in precision.txt)

OUTPUT}
}
\value{
number of digits after the comman for the variable
of interest. If the variable is not found in precision.txt
FALSE is returned instead
}
\description{
#############################################################
GET PRECISION
DESCRIPTION
returns the precision (number of digits) after the comma for
the variable of interest
PARAMETERS
}
