% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadagno.r
\name{plot_set}
\alias{plot_set}
\title{plot_set}
\usage{
plot_set(x, set = "abio", lwd = 1, Rbrewer_colorname = "RdBu", ...)
}
\arguments{
\item{x}{cleaned .TOB dataframe}

\item{set}{set of variables of interest (vector).
Otherwise,
"abio", to put Temp, Cond, DO_mg and Turb into a single plot
"bio", PAR, Chl_A and BGAPC into single plot.}

\item{...}{additional parameter to plot()

OUTPUT
nothing}
}
\description{
DESCRIPTION
plot a set of measurements of interest against depth, using
moving average. All variables are displayed in the same plot
}
\details{
PARAMETERS
}
