% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadagno.r
\name{plot_var}
\alias{plot_var}
\title{plot_var}
\usage{
plot_var(var, depth, xlab = "var", linecolor = "cyan4", lwd = 2, ...)
}
\arguments{
\item{var}{variable (vector)}

\item{depth}{depth values (vector)}

\item{...}{additional parameter to plot()

OUTPUT
nothing}
}
\description{
DESCRIPTION
plot the profile of the specified variable against depth
using simple moving average (?TTR::SMA)
}
\details{
PARAMETERS
}
