% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadagno.r
\name{read_TOB}
\alias{read_TOB}
\title{read_TOB}
\usage{
read_TOB(fname, showunits = FALSE, outfile = FALSE, DEBUG = FALSE)
}
\arguments{
\item{fname}{file path/name to open for reading}

\item{showunits}{column names with parameter units}

\item{outfile}{name of the output file. If false,
then a default filename is used}

\item{DEBUG}{code maintenance only

OUTPUT}
}
\value{
a data frame
}
\description{
DESCRIPTION
this function reads .TOB data file and returns a dataframe
}
\details{
PARAMETERS
}
