% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadagno.r
\name{plot_set}
\alias{plot_set}
\title{plot_set}
\usage{
plot_set(
  x,
  BL = "abio",
  lwd = 1,
  Rbrewer_colorname = "RdBu",
  zoomIn = FALSE,
  ...
)
}
\arguments{
\item{x}{cleaned .TOB dataframe}

\item{...}{additional parameter to plot()

OUTPUT
nothing}

\item{set}{set of variables of interest (vector).
Otherwise,
"abio", displays Temp, Cond, DO_mg and Turb in a single plot
"bio", PAR, Chl_A and BGAPC into single plot.}
}
\description{
DESCRIPTION
plot a set of measurements of interest against depth, using
moving average. All variables are displayed in the same plot
}
\details{
PARAMETERS
}
